display vpc in the network interfaces column of the config table
